【A】添加消息列表消息间切换用例，修改工具的操作类接口

+++ ApkAutoTest/app/src/main/java/com/tencent/apk_auto_test/util/Global.java
@@ -1,6 +1,6 @@
 package com.tencent.apk_auto_test.util;
 
 public class Global {
-    public static int SCREEN_WIDTH = 400;
-    public static int SCREEN_HEIGHT = 800;
+    public static double SCREEN_WIDTH = 400;
+    public static double SCREEN_HEIGHT = 800;
 }
